Move NaFlexCollate with dataset, remove stand alone collate_fn and remove redundancy
